trying to standartize a database sql doesn't support unsigned types by default, so im need to change everything there's still a lot of work around for site, but at least new api is already done and this api is freaking awesome

+++ src/beatmaps/beatmap.hh
@@ -56,7 +56,7 @@ namespace shiro::beatmaps {
 
         int32_t max_combo = 0;
         int32_t hit_length = 0;
-        int32_t bpm = 0;
+        int64_t bpm = 0;
 
         int32_t count_normal = 0;
         int32_t count_slider = 0;
